requirements.txt
amqp==2.6.1
azure-core==1.8.0
azure-cosmos==4.1.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
funcsigs==1.0.2
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
mock==4.0.2
mongolog==0.1.2
motor==2.1.0
pbr==5.4.5
pika==1.1.0
pkg-resources==0.0.0
pymongo==3.11.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
